+++ PJE3JX/app/src/main/res/layout-land/activity_fragment.xml
@@ -1,0 +1,24 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="horizontal"
+    android:baselineAligned="false"
+    tools:context=".DemoActivity"
+    >
+
+
+<fragment
+        android:id="@+id/left_fragment"
+        android:name="fr.univ_lille1.pje.pje3jx.BookListFragment"
+        android:layout_width="0dip"
+        android:layout_height="fill_parent"
+        android:layout_weight="2"
+        />
+    <FrameLayout
+        android:id="@+id/details_layout"
+        android:layout_width="0dip"
+        android:layout_height="fill_parent"
+        android:layout_weight="3"
+        ></FrameLayout>
+</LinearLayout>
