+++ PJE3JX/app/src/main/res/layout/activity_fragment.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <fragment
+        android:id="@+id/fragment_list_books"
+        android:name="fr.univ_lille1.pje.pje3jx.BookListFragment"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent">
+    </fragment>
+
+</LinearLayout>
