+++ PJE3JX/app/src/main/res/layout/fragment_book_list.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="fr.univ_lille1.pje.pje3jx.BookListActivity">
+    tools:context="fr.univ_lille1.pje.pje3jx.BookListFragment">
 
     <ListView
         android:id="@+id/listView"
